Implement resolving of collisions at different resolutions

+++ src/game/physics/manifold.h
@@ -1,4 +1,4 @@
 struct Manifold {
-    int norm_x = 0, norm_y = 0;
+    double norm_x = 0, norm_y = 0;
     double cost = -1;
 };
